requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
crispy-bootstrap5==2025.6
cryptography==45.0.6
distlib==0.3.9
distro==1.9.0
dj-database-url==3.0.1
Django==5.1.1
django-crispy-forms==2.4
django-debug-toolbar==5.2.0
django-otp==1.6.0
django-redis==6.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
Faker==37.4.0
filelock==3.16.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
joblib==1.5.1
msgpack==1.1.1
nltk==3.9.1
numpy==2.1.3
openai==1.99.9
outcome==1.3.0.post0
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.0.0
platformdirs==4.3.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pypdfium2==4.30.0
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.1.0
redis==6.2.0
regex==2025.7.34
selenium==4.35.0
setuptools==80.9.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.1
urllib3==2.5.0
virtualenv==20.28.0
virtualenvwrapper-win==1.2.7
websocket-client==1.8.0
wheel==0.45.1
whitenoise==6.9.0
wsproto==1.2.0
